Moved progress_bar.py to the vcstools module
