+++ vcstools/progress_bar.py
